+++ WebApp/Models/Company.cs
@@ -7,7 +7,7 @@ namespace WebApp.Models
 {
     public class Company
     {
-        public int Id { get; set; }
+        public short Id { get; set; }
         public string Name { get; set; }
         public string LogoCssClass { get; set; }
         public string MasterBookingPageByTime { get; set; }
